Quiz implemented
Add QuizManager prefab as a child of the level canvas and the quiz will appear on the screen and work. Adjust questions in the inspector for the QuizManager script. Questions should be entered in the QuestionsInput[] field and formatted as follows: question text\correct answer\ incorrect answer1\ incorrect answer2\incorrect answer3